DOC-573: add mount PVC topic (#330)
For https://hazelcast.atlassian.net/browse/DOC-573

As part of the Diagnostic Logging work, we need to document how to mount
one or more Persistent Volume Claims (PVCs) directly to a Hazelcast
cluster.

Sergii suggests this goes into a new topic under Configuration
properties.

The work to add the fields to the CRD has been done in PR:
https://github.com/hazelcast/hazelcast-platform-operator/pull/1432